requirements.txt
pandas~=2.1.1
numpy~=1.26.1
python-dateutil~=2.8.2